requirements.txt
python-dateutil>=2.8.2
requests>=2.27.1
jsonschema~=3.2.0
httpx
pyyaml
biothings-explorer>=1.1.0
bmt>=0.8.4
reasoner-validator>=2.2.1
pytest>=7.1.1
pytest-asyncio
pytest-harvest
kgx
